Chain up Zalgo's cousin the exhausted stream.
The base class now wraps all incoming data in a stream (if necessary)
and pipes immediately to a pass-through stream before handing the task
off to individual clients. This ensures that all the clients get piped
to properly in the first tick, so no one client can race ahead & snag
all the data before others are read.

Removed repeated stream-wrapping code from the implementations of
_writeFileCore. (Manta continues to have its own standalone implementation.)